+++ scripts/macros/homebrew/criticalFumble/examples.js
@@ -1,7 +1,7 @@
 import {chris} from '../../../helperFunctions.js';
 import {queue} from '../../../utility/queue.js';
 async function maxDamage({speaker, actor, token, character, item, args, scope, workflow}) {
-    Hooks.once('midi-qol.DamageRollComplete', async (workflow) => {
+    Hooks.once('midi-qol.preDamageRollComplete', async (workflow) => {
         let queueSetup = await queue.setup(workflow.item.uuid, 'maxDamage', 450);
         if (!queueSetup) return;
         await Promise.all(workflow.damageRolls.map(async (damageRoll, i, arr) => {
@@ -12,11 +12,11 @@ async function maxDamage({speaker, actor, token, character, item, args, scope, w
     });
 }
 async function multiplyDamage(multiplier) {
-    Hooks.once('midi-qol.DamageRollComplete', async (workflow) => {
+    Hooks.once('midi-qol.preDamageRollComplete', async (workflow) => {
         let queueSetup = await queue.setup(workflow.item.uuid, 'multiplyDamage', 450);
         if (!queueSetup) return;
         await Promise.all(workflow.damageRolls.map(async (damageRoll, i, arr) => {
-            arr[i] = await chris.damageRoll(workflow, '(' + damageRoll._formula + ') * ' + multiplier, undefined, true);
+            arr[i] = await chris.damageRoll(workflow, '((' + damageRoll._formula + ') * ' + multiplier + ')', undefined, true);
         }));
         await workflow.setDamageRolls(workflow.damageRolls);
         queue.remove(workflow.item.uuid);